+++ controller/editorcontroller.php
@@ -35,6 +35,8 @@ use OCP\IL10N;
 use OCP\ILogger;
 use OCP\IRequest;
 use OCP\ISession;
+use OCP\ITags;
+use OCP\ITagManager;
 use OCP\IURLGenerator;
 use OCP\IUser;
 use OCP\IUserManager;
@@ -135,6 +137,13 @@ class EditorController extends Controller {
     private $shareManager;
 
     /**
+     * Tag manager
+     *
+     * @var ITagManager
+    */
+    private $tagManager;
+
+    /**
      * Mobile regex from https://github.com/ONLYOFFICE/CommunityServer/blob/v9.1.1/web/studio/ASC.Web.Studio/web.appsettings.config#L35
      */
     const USER_AGENT_MOBILE = "/android|avantgo|playbook|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(hone|od|ad)|iris|kindle|lge |maemo|midp|mmp|opera m(ob|in)i|palm( os)?|phone|p(ixi|re)\\/|plucker|pocket|psp|symbian|treo|up\\.(browser|link)|vodafone|wap|windows (ce|phone)|xda|xiino/i";
@@ -152,6 +161,7 @@ class EditorController extends Controller {
      * @param Crypt $crypt - hash generator
      * @param IManager $shareManager - Share manager
      * @param ISession $ISession - Session
+     * @param ITagManager $tagManager - Tag manager
      */
     public function __construct($AppName,
                                     IRequest $request,
@@ -164,7 +174,8 @@ class EditorController extends Controller {
                                     AppConfig $config,
                                     Crypt $crypt,
                                     IManager $shareManager,
-                                    ISession $session
+                                    ISession $session,
+                                    ITagManager $tagManager
                                     ) {
         parent::__construct($AppName, $request);
 
@@ -177,6 +188,7 @@ class EditorController extends Controller {
         $this->config = $config;
         $this->crypt = $crypt;
         $this->shareManager = $shareManager;
+        $this->tagManager = $tagManager;
 
         if (\OC::$server->getAppManager()->isInstalled("files_versions")) {
             try {
@@ -1032,19 +1044,6 @@ class EditorController extends Controller {
     }
 
     /**
-     * Get template loader Onlyoffice
-     *
-     * @return TemplateResponse
-     *
-     * @NoAdminRequired
-     * @NoCSRFRequired
-     * @PublicPage
-     */
-    public function loader() {
-        return new TemplateResponse($this->appName, "loader", [], "plain");
-    }
-
-    /**
      * Collecting the file parameters for the document service
      *
      * @param integer $fileId - file identifier
@@ -1267,6 +1266,9 @@ class EditorController extends Controller {
 
                 $params["editorConfig"]["templates"] = $templates;
             }
+
+            $params["document"]["info"]["favorite"] = $this->isFavorite($fileId);
+            $params["_file_path"] = $userFolder->getRelativePath($file->getPath());
         }
 
         if ($folderLink !== null) {
@@ -1617,6 +1619,22 @@ class EditorController extends Controller {
                     return $watermarkText;
                 }
             }
+        }
+
+        return false;
+    }
+
+    /**
+     * Check file favorite
+     *
+     * @param integer $fileId - file identifier
+     *
+     * @return bool
+     */
+    private function isFavorite($fileId) {
+        $currentTags = $this->tagManager->load("files")->getTagsForObjects([$fileId]);
+        if ($currentTags) {
+            return in_array(ITags::TAG_FAVORITE, $currentTags[$fileId]);
         }
 
         return false;