+++ js/editor.js
@@ -156,6 +156,11 @@
                         config.events.onRequestCompareFile = OCA.Onlyoffice.onRequestCompareFile;
                         config.events.onRequestUsers = OCA.Onlyoffice.onRequestUsers;
                         config.events.onRequestSendNotify = OCA.Onlyoffice.onRequestSendNotify;
+                        config.events.onMetaChange = OCA.Onlyoffice.onMetaChange;
+
+                        if (!OCA.Onlyoffice.filePath) {
+                            OCA.Onlyoffice.filePath = config._file_path;
+                        }
                     }
 
                     if (OCA.Onlyoffice.directEditor || OCA.Onlyoffice.inframe) {
@@ -499,6 +504,23 @@
                 OCP.Toast.success(response.message);
             });
     };
+
+    OCA.Onlyoffice.onMetaChange = function (event) {
+        if (event.data.favorite !== undefined) {
+            $.ajax({
+                url: OC.generateUrl("apps/files/api/v1/files" + OC.encodePath(OCA.Onlyoffice.filePath)),
+                type: "post",
+                data: JSON.stringify({
+                    tags: event.data.favorite ? [OC.TAG_FAVORITE] : []
+                }),
+                contentType: "application/json",
+                dataType: "json",
+                success: function(){
+                    OCA.Onlyoffice.docEditor.setFavorite(event.data.favorite);
+                }
+            });
+        }
+    }
 
     OCA.Onlyoffice.InitEditor();